# src/main/resources/static/swagger.yml
openapi: 3.0.3
info:
  title: Online Shop
  description: |-
    Aplicación que se encarga de habilitar endpoints para el consumo REST de la base de datos en la tienda online.
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: http://localhost:8080/api/v1
    description: Local server
...
paths:
...
  /auth/register:
    post:
      tags:
        - auth
      summary: Register a user
      description: Register a user and returns an access token and refresh token.
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRequest'
        required: true
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Invalid input
        '422':
          description: Validation exception
  /auth/login:
    post:
      tags:
        - auth
      summary: Log in a user
      description: Authenticates a user and returns an access token and refresh token.
      operationId: loginUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResponse'
        '400':
          description: Invalid input
        '401':
          description: Authentication failed
  /auth/verify/{token}:
    post:
      tags:
        - auth
      summary: Verificar la cuenta de un usuario usando el token de verificación
      description: Verifica la cuenta del usuario utilizando el token de verificación proporcionado.
      operationId: verifyUser
      parameters:
        - name: token
          in: path
          description: El token de verificación para verificar la cuenta
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Cuenta verificada exitosamente
          content:
            application/json:
              schema:
                type: string
                example: "Cuenta verificada exitosamente."
        '400':
          description: Token de verificación inválido
          content:
            application/json:
              schema:
                type: string
                example: "Token de verificación inválido."
        '404':
          description: Token de verificación no encontrado
          content:
            application/json:
              schema:
                type: string
                example: "Token de verificación no encontrado."
  /auth/authenticated-user:
    get:
      tags:
        - auth
      summary: Obtener la información del usuario autenticado
      description: Recupera la información del usuario autenticado.
      operationId: getAuthenticatedUser
      responses:
        '200':
          description: Usuario autenticado encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                type: string
                example: "No se encontró ningún usuario autenticado."
  /auth/password-reset/request:
    post:
      tags:
        - auth
      summary: Envia el link de resetear la contraseña
      description: Envia el codigo de resetear la contraseña al email del usuario
      operationId: passwordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  description: The email address of the user.
                  example: user@example.com
      responses:
        '200':
          description: Password reset link sent successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailResponse'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                type: string
                example: "No se encontró ningún usuario autenticado."
  /auth/password-reset/confirm:
    patch:
      tags:
        - auth
      summary: Resetea la contrasela utilizando el token
      description: Si el token es valido, resetea la contraseña
      operationId: passwordResetConfirm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                verification-token:
                  type: string
                  description: The password reset token sent to the user's email.
                  example: dummytoken123
                newPassword:
                  type: string
                  format: password
                  description: The new password for the user.
                  example: P@ssw0rd123
      responses:
        '200':
          description: Password reset successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '400':
          description: Invalid or expired token.
          content:
            application/json:
              schema:
                type: string
                example: "No se encontró ningún usuario autenticado."
...
  /shop/carts:
    get:
      tags:
        - cart
      summary: Obtener todos los carritos de compra
      description: Devuelve una lista de todos los carritos de compra.
      operationId: getAllCarts
      responses:
        '200':
          description: Operación exitosa, devuelve una lista de carritos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CartResponse'
        '400':
          description: Solicitud incorrecta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /shop/cart/{cartId}:
    get:
      tags:
        - cart
      summary: Obtener carrito de la compra
      description: Devuelve el carrito de la compra
      operationId: getCart
      parameters:
        - name: cartId
          in: path
          description: 'Id del carrito a buscar '
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '404':
          description: Carrito no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /shop/create-cart:
    post:
      tags:
        - cart
      summary: Crear un nuevo carrito de la compra
      description: Crea un nuevo carrito de la compra y devuelve el carrito creado
      operationId: createCart
      requestBody:
        description: 'Datos para la creación del carrito'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartRequest'
      responses:
        '200':
          description: Carrito creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '400':
          description: Solicitud incorrecta
  /shop/cart/{cartId}/add-product:
    post:
      tags:
        - cart
      summary: Añadir producto al carrito de la compra
      description: Devuelve el carrito de la compra actualizado
      operationId: addCartProduct
      parameters:
        - name: cartId
          in: path
          description: 'Id del carrito a buscar'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: 'RequestBody de los detalles del carrito'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartDetailsRequest'
      responses:
        '200':
          description: Operación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartDetailsResponse'
        '404':
          description: El carrito no ha podido ser creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /shop/cart/{cartId}/update-cart:
    put:
      tags:
        - cart
      summary: Actualizar información del carrito de la compra
      description: Devuelve el carrito de la compra actualizado
      operationId: updateCart
      parameters:
        - name: cartId
          in: path
          description: 'Id del carrito a actualizar'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: 'RequestBody con la información a actualizar del carrito'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartRequest'
      responses:
        '200':
          description: Operación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '404':
          description: Carrito no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Solicitud incorrecta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /shop/cart/delete-cart/{cartId}:
    delete:
      tags:
        - cart
      summary: Eliminar carrito de la compra
      description: Elimina el carrito de la compra identificado por el `cartId`.
      operationId: deleteCart
      parameters:
        - name: cartId
          in: path
          description: 'ID del carrito a eliminar'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Carrito eliminado exitosamente
          content:
            text/plain:
              schema:
                type: string
                example: "El carrito ha sido eliminado con éxito"
        '404':
          description: Carrito no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Solicitud incorrecta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /shop/cart/{cartId}/remove-product/{cartDetailsId}:
    delete:
      tags:
        - cart
      summary: Eliminar un producto del carrito de la compra
      description: Elimina un producto del carrito de la compra
      operationId: removeCartProduct
      parameters:
        - name: cartId
          in: path
          description: 'Id del carrito a buscar'
          required: true
          schema:
            type: integer
            format: int64
        - name: cartDetailsId
          in: path
          description: 'Id de los detalles del carrito'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Operación exitosa
          content:
            text/plain:
              schema:
                type: string
                example: "El producto ha sido eliminado del carrito con éxito"
...
components:
  schemas:
...
    EmailResponse:
      type: object
      properties:
        message:
          type: string
    CartRequest:
      type: object
      properties:
        cartDetails:
          type: array
          items:
            $ref: '#/components/schemas/CartDetailsRequest'
      required:
        - id
        - date
        - deliveryAddress
        - paymentMethod
    CartResponse:
      type: object
      properties:
        userId:
          type: integer
          format: int64
          example: 1
        date:
          type: string
          format: date-time
          example: "2024-05-26T12:30:00Z"
        cartDetails:
          type: array
          items:
            $ref: '#/components/schemas/CartDetailsResponse'
    CartDetailsRequest:
      type: object
      properties:
        productId:
          type: integer
          format: int64
          example: 2
        quantity:
          type: integer
          example: 1
    CartDetailsResponse:
      type: object
      properties:
        productId:
          type: integer
          format: int64
          example: 2
        price:
          type: number
          example: 29.99
        quantity:
          type: integer
          example: 1
...
    CustomerRequest:
      type: object
      properties:
        name:
          type: string
          example: John
        password:
          type: string
          example: '12345'
        surname:
          type: string
          example: James
        surname2:
          type: string
          example: James
        address:
          type: string
          example: Calle nombre n3 1b
        city:
          type: string
          example: Chiclana de la Frontera
        province:
          type: string
          example: Cadiz
        region:
          type: string
          example: España
        postalCode:
          type: string
          example: "11130"
        email:
          type: string
          example: john@email.com
        phone:
          type: string
          example: '12345'
        status:
          type: boolean
          description: Estado del usuario
          example: true
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          example: john@email.com
        password:
          type: string
          example: '12345'
    CustomerResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        username:
          type: string
          example: theUser
        password:
          type: string
          example: '12345'
        name:
          type: string
          example: John
        surname:
          type: string
          example: James
        surname2:
          type: string
          example: James
        address:
          type: string
          example: Calle nombre n3 1b
        city:
          type: string
          example: Chiclana de la Frontera
        province:
          type: string
          example: Cadiz
        region:
          type: string
          example: España
        postalCode:
          type: string
          example: "11130"
        email:
          type: string
          example: john@email.com
        phone:
          type: string
          example: '12345'
        status:
          type: boolean
          description: Estado del usuario
          example: true
        rol:
          type: string
          format: enum
            - USER
            - ADMIN
          description: Rol de usuario
          example: USER
    AuthenticationResponse:
      type: object
      properties:
        accessToken:
          type: string
          example: eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQSflwTmsuHo_iij07PpRZsI7UlFz06OZtToOZfpC9q6U
        refreshToken:
          type: string
          example: dGhpcy1yZWZyZXNoLXRva2Vu
        message:
          type: string
          example: Authentication successful
    RegisterResponse:
      type: object
      properties:
        accessToken:
          type: string
          example: null
        refreshToken:
          type: string
          example: null
        message:
          type: string
          example: Usuario registrado correctamente. Por favor revisa tu correo electrónico para verificar la cuenta.
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
          example: 404
          description: "Código de estado HTTP que indica el tipo de error."
        message:
          type: string
          example: "Cart not found"
          description: "Mensaje descriptivo del error que proporciona más detalles."
        details:
          type: string
          example: "The cart with the given ID does not exist."
          description: "Información adicional o detalles sobre el error (opcional)."